Move routes to habitability dir
