+++ frontend/lib/laletterbuilder/route-info.ts
@@ -1,10 +1,6 @@
 import { createDevRouteInfo } from "../dev/route-info";
-import {
-  createHtmlEmailStaticPageRouteInfo,
-  createLetterStaticPageRouteInfo,
-} from "../static-page/routes";
 import { ROUTE_PREFIX, createRoutesForSite } from "../util/route-util";
-import { createLaLetterBuilderRouteInfo } from "./letter-builder/route-info";
+import { createHabitabilityRouteInfo } from "./letter-builder/habitability/route-info";
 
 function createLocalizedRouteInfo(prefix: string) {
   return {
@@ -14,30 +10,16 @@ function createLocalizedRouteInfo(prefix: string) {
     /** The home page. */
     home: `${prefix}/`,
 
-    /** The letter builder */
-    letter: createLaLetterBuilderRouteInfo(`${prefix}/letter`),
+    chooseLetter: `${prefix}/choose-letter`,
 
-    /** The letter content for the user's own data (HTML and PDF versions). */
-    letterContent: createLetterStaticPageRouteInfo(`${prefix}/letter`),
+    /** Habitability Letter flow */
+    habitability: createHabitabilityRouteInfo(`${prefix}/habitability`),
 
     /** The logout page. */
     logout: `${prefix}/logout`,
 
     /** The about page. */
     about: `${prefix}/about`,
-
-    /** The email to the user's landlord. */
-    letterEmail: `${prefix}/letter-email.txt`,
-
-    /** The email to the user w/ a copy of the letter. */
-    letterEmailToUser: createHtmlEmailStaticPageRouteInfo(
-      `${prefix}/letter-email-to-user`
-    ),
-
-    /** The sample letter content (HTML and PDF versions). */
-    sampleLetterContent: createLetterStaticPageRouteInfo(
-      `${prefix}/sample-letter`
-    ),
   };
 }
 
@@ -53,7 +35,7 @@ export const LaLetterBuilderRouteInfo = createRoutesForSite(
 );
 
 export const getLaLetterBuilderJumpToTopOfPageRoutes = () => [
-  LaLetterBuilderRouteInfo.locale.letter.confirmation,
+  LaLetterBuilderRouteInfo.locale.habitability.confirmation, // add a ref to all confirmation pages for all 4 letters
   ...getLaLetterBuilderRoutesForPrimaryPages(),
 ];
