+++ frontend/lib/laletterbuilder/routes.tsx
@@ -8,14 +8,9 @@ import { AlternativeLogoutPage } from "../pages/logout-alt-page";
 import { NotFound } from "../pages/not-found";
 import { LaLetterBuilderAboutPage } from "./about";
 import { LaLetterBuilderHomepage } from "./homepage";
-import { LaLetterBuilderRoutes } from "./letter-builder/routes";
 import { LaLetterBuilderRouteInfo as Routes } from "./route-info";
-import { createLetterStaticPageRoutes } from "../static-page/routes";
-import {
-  HabitabilityLetterEmailToLandlordForUserStaticPage,
-  HabitabilityLetterForUserStaticPage,
-  HabitabilitySampleLetterSamplePage,
-} from "./letter-builder/habitability-letter-content";
+import HabitabilityRoutes from "./letter-builder/habitability/routes";
+import { LaLetterBuilderChooseLetterStep } from "./letter-builder/choose-letter";
 
 const LoadableDevRoutes = loadable(
   () => friendlyLoad(import("../dev/routes")),
@@ -50,22 +45,13 @@ export const LaLetterBuilderRouteComponent: React.FC<RouteComponentProps> = (
         component={AlternativeLogoutPage}
       />
       <Route
-        path={Routes.locale.letter.prefix}
-        component={LaLetterBuilderRoutes}
+        path={Routes.locale.habitability.prefix}
+        component={HabitabilityRoutes}
       />
-      {createLetterStaticPageRoutes(
-        Routes.locale.letterContent,
-        HabitabilityLetterForUserStaticPage
-      )}
       <Route
-        path={Routes.locale.letterEmail}
-        exact
-        component={HabitabilityLetterEmailToLandlordForUserStaticPage}
+        path={Routes.locale.chooseLetter}
+        component={LaLetterBuilderChooseLetterStep}
       />
-      {createLetterStaticPageRoutes(
-        Routes.locale.sampleLetterContent,
-        HabitabilitySampleLetterSamplePage
-      )}
       <Route path={Routes.dev.prefix} component={LoadableDevRoutes} />
       <Route component={NotFound} />
     </Switch>
